refactor new or edit invoice product form. add validator for invoice product.

+++ PriceBookClassLibrary/InvoiceProductModel.cs
@@ -9,7 +9,7 @@ namespace PriceBookClassLibrary
         public int ProductId { get; set; }
         public string ProductLinkName { get; set; }
         public string ProductName { get; set; }
-        public int Weight {get; set;}
+        public decimal Weight {get; set;}
         public int Quantity {get; set;}
         public decimal TotalPrice {get; set;}
         public string Sale { get; set; }
